requirements.txt
openai>=1.3.0
google-generativeai==0.4.0
google-cloud-aiplatform>=1.26.1
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.1
requests>=2.31.0